requirements.txt
google-api-python-client==1.11.0
google-auth-httplib2==0.0.4
python-telegram-bot==12.8
google-auth-oauthlib==0.4.1
pyyaml==5.3.1